Rename main module! Fix the simple (monitor.py) script and update `__init__.py` import for ble2lsl.
