+++ scripts/monitor.py
@@ -1,12 +1,12 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""Binary SVM classifier of concentrating vs. relaxating states from EEG data."""
 
-from musemonitor import online
-import gatt2lsl
+"""
+
+"""
+
+from wizardhat import acquire
+import ble2lsl
 
 if __name__ == '__main__':
-    outlet_streamer = gatt2lsl.LSLOutletStreamer()
-	#outlet_streamer = gatt2lsl.LSLOutletDummy() #Let's add default arg so people can run either dummy or real without commenting
-	print('Connected')
-    streamer = online.LSLStreamer()
+    outlet_streamer = ble2lsl.LSLOutletStreamer()
+    # outlet_streamer = ble2lsl.LSLOutletDummy()
+    streamer = acquire.LSLStreamer()
